Push with working numbers.
CSV Export expected tomorrow morning

+++ AClient.cs
@@ -41,9 +41,9 @@ namespace GrenciCPA
         private string preferred;
         private int jobID;
         private string dateSent;
-        private float totalBill;
+        private decimal totalBill;
         private string servName;
-        private float paymentAmount;
+        private decimal paymentAmount;
 
 
 
@@ -71,8 +71,8 @@ namespace GrenciCPA
         public string Preferred { get => preferred; set => preferred = value; }
         public int JobID { get => jobID; set => jobID = value; }
         public string DateSent { get => dateSent; set => dateSent = value; }
-        public float TotalBill { get => totalBill; set => totalBill = value; }
+        public decimal TotalBill { get => totalBill; set => totalBill = value; }
         public string ServName { get => servName; set => servName = value; }
-        public float PaymentAmount { get => paymentAmount; set => paymentAmount = value; }
+        public decimal PaymentAmount { get => paymentAmount; set => paymentAmount = value; }
     }
 }
